Use `assertEventually` to make tests more reliable when latency is bad
I'm currently in Brazil and experiencing poor latency, which is making the tests
fail the way that they're written. I was able to get them passing by using
`assertEventually` to repeatedly check until the condition passes, failing once
a max timeout is reached.

(TODO: Remove this inlined version of `assertEventually` and pull it in from
mongochangestream-testing instead.)

Results:

 ? src/convertSchema.test.ts (10)
 ? src/crate.test.ts (7)
 ? src/syncData.test.ts (3) 41630ms
 ? src/util.test.ts (3)

 Test Files  4 passed (4)
      Tests  23 passed (23)
   Start at  14:01:45
   Duration  42.14s (transform 121ms, setup 0ms, collect 610ms, tests 41.65s, environment 0ms, prepare 189ms)